requirements.txt
# Figma to XAML Converter V2 依赖
jinja2>=3.1.0
pyyaml>=6.0
